refactor: centralize MVP card types
