+++ src/components/game/CardCollectionTabloid.tsx
@@ -6,7 +6,7 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@
 import { Progress } from '@/components/ui/progress';
 import { Button } from '@/components/ui/button';
 import { useCardCollection } from '@/hooks/useCardCollection';
-import type { GameCard } from '@/types/cardTypes';
+import type { GameCard } from '@/rules/mvp';
 
 interface CardCollectionTabloidProps {
   open: boolean;
