+++ src/mvp/index.ts
@@ -1,5 +1,5 @@
 export const USE_MVP_RULES = true;
 
-export * from './types';
+export * from '@/rules/mvp';
 export * from './validator';
 export * from './engine';
